remote-setup/requirements.txt
# packages required for all environments
aiohttp-jinja2==1.4.2
aiopg==1.1.0
asyncpg==0.21.0
cchardet==2.1.7
psycopg2==2.8.6
SQLAlchemy==1.3.22
ujson==4.0.1
uvloop==0.14.0
